# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: German Flashcard App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: flashcards
    description: German flashcard operations
  - name: progress
    description: User study progress
  - name: workspaces
    description: Per-user dashboards (German→X language pairs)
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /flashcards:
    get:
      operationId: listFlashcards
      tags: [flashcards]
      summary: List flashcards with optional level filter
      parameters:
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [A1, A2, B1, B2, C1]
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of flashcards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardList"

  /flashcards/generate:
    post:
      operationId: generateFlashcards
      tags: [flashcards]
      summary: Generate new flashcards using AI for a specific level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardGenerateInput"
      responses:
        "201":
          description: Generated flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"

  /flashcards/stats:
    get:
      operationId: getFlashcardStats
      tags: [flashcards]
      summary: Get study stats grouped by level
      responses:
        "200":
          description: Stats by level
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LevelStats"

  /flashcards/daily:
    get:
      operationId: getDailyFlashcards
      tags: [flashcards]
      summary: Get today's recommended flashcard set
      parameters:
        - name: level
          in: query
          required: false
          schema:
            type: string
            enum: [A1, A2, B1, B2, C1]
      responses:
        "200":
          description: Daily flashcard set
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"

  /flashcards/{id}:
    get:
      operationId: getFlashcard
      tags: [flashcards]
      summary: Get a single flashcard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /flashcards/{id}/progress:
    patch:
      operationId: updateFlashcardProgress
      tags: [progress]
      summary: Mark a flashcard as known or unknown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /me/workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List the caller's workspaces (default + user-created)
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace (max 2 per user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Workspace limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /me/workspaces/{id}/switch:
    post:
      operationId: switchWorkspace
      tags: [workspaces]
      summary: Set the active workspace for the caller
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceSwitchResult"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /me/workspaces/{id}:
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete a user-created workspace
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: string
                required:
                  - deleted
        "400":
          description: Cannot delete the default workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Workspace not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Flashcard:
      type: object
      properties:
        id:
          type: integer
        word:
          type: string
          description: German word with article e.g. "der Hund"
        article:
          type: ["string", "null"]
          description: der / die / das / null for verbs/adjectives
        baseWord:
          type: string
          description: The word without article
        level:
          type: string
          enum: [A1, A2, B1, B2, C1]
        category:
          type: string
        englishTranslation:
          type: string
        arabicTranslation:
          type: string
        exampleSentenceDe:
          type: string
        exampleSentenceEn:
          type: string
        exampleSentenceAr:
          type: string
        imageUrl:
          type: ["string", "null"]
        known:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - word
        - article
        - baseWord
        - level
        - category
        - englishTranslation
        - arabicTranslation
        - exampleSentenceDe
        - exampleSentenceEn
        - exampleSentenceAr
        - imageUrl
        - known
        - createdAt

    FlashcardList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - offset
        - limit

    FlashcardGenerateInput:
      type: object
      properties:
        level:
          type: string
          enum: [A1, A2, B1, B2, C1]
        category:
          type: string
        count:
          type: integer
          default: 10
      required:
        - level

    LevelStats:
      type: object
      properties:
        level:
          type: string
        total:
          type: integer
        known:
          type: integer
        unknown:
          type: integer
        percentage:
          type: integer
      required:
        - level
        - total
        - known
        - unknown
        - percentage

    ProgressUpdate:
      type: object
      properties:
        known:
          type: boolean
      required:
        - known

    Workspace:
      type: object
      properties:
        id:
          type: string
          description: The string 'default' for the virtual default workspace, otherwise a UUID
        name:
          type: string
        secondaryLanguage:
          type: string
          description: ISO-ish code (AR, EN, ES, FR, IT, TR)
        isDefault:
          type: boolean
        isCurrent:
          type: boolean
      required:
        - id
        - name
        - secondaryLanguage
        - isDefault
        - isCurrent

    WorkspaceList:
      type: object
      properties:
        currentId:
          type: ["string", "null"]
          description: Null when the user is in the default workspace
        max:
          type: integer
          description: Max number of additional user-created workspaces (3 total dashboards)
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/Workspace"
      required:
        - currentId
        - max
        - workspaces

    WorkspaceCreateInput:
      type: object
      properties:
        secondaryLanguage:
          type: string
          enum: [EN, ES, FR, IT, TR]
        name:
          type: string
          description: Optional; defaults to language name
      required:
        - secondaryLanguage

    WorkspaceSwitchResult:
      type: object
      properties:
        currentId:
          type: ["string", "null"]
      required:
        - currentId

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error