# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * German Flashcard App API
 * OpenAPI spec version: 0.1.0
 */

export interface Workspace {
  /** The string 'default' for the virtual default workspace, otherwise a UUID */
  id: string;
  name: string;
  /** ISO-ish code (AR, EN, ES, FR, IT, TR) */
  secondaryLanguage: string;
  isDefault: boolean;
  isCurrent: boolean;
}