# lib/api-zod/src/generated/types/workspaceSwitchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * German Flashcard App API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspaceSwitchResult {
  /** @nullable */
  currentId: string | null;
}